Fix failing tests and circular import. Have the CompetitionSpecification inherit from BenchmarkSpecification.
