+++ polaris/evaluate/__init__.py
@@ -1,17 +1,16 @@
+from polaris.evaluate._metadata import ResultsMetadata
 from polaris.evaluate._metric import Metric, MetricInfo
 from polaris.evaluate._predictions import BenchmarkPredictions, CompetitionPredictions
 from polaris.evaluate._results import (
     BenchmarkResults,
     CompetitionResults,
     EvaluationResult,
-    ResultsMetadata,
 )
 from polaris.evaluate.utils import evaluate_benchmark
 
 __all__ = [
     "Metric",
     "MetricInfo",
-    "ResultsMetadata",
     "EvaluationResult",
     "BenchmarkResults",
     "CompetitionResults",
