+++ tests/conftest.py
@@ -319,7 +319,7 @@ def test_multi_task_benchmark_clf(test_dataset, classification_metrics):
 
 
 @pytest.fixture(scope="function")
-def test_competition(zarr_archive, test_org_owner, regression_metrics):
+def test_competition(zarr_archive, test_org_owner, regression_metrics, test_dataset_v2):
     train_indices = list(range(90))
     test_indices = list(range(90, 100))
     competition = CompetitionSpecification(
@@ -332,20 +332,19 @@ def test_competition(zarr_archive, test_org_owner, regression_metrics):
         metrics=regression_metrics,
         main_metric="mean_absolute_error",
         split=(train_indices, test_indices),
-        target_cols=["expt"],
-        input_cols=["smiles"],
+        input_cols=["A"],
+        target_cols=["B"],
         readme="Testing specification",
-        start_time="2025-01-13T21:59:38Z",
-        end_time="2025-01-20T21:59:38Z",
-        n_test_sets=1,
-        n_test_datapoints={"test": 10},
-        n_classes={"test": 0},
         # Dataset attributes
         source="https://www.example.com",
         annotations={"A": ColumnAnnotation(user_attributes={"unit": "kcal/mol"})},
         license="CC-BY-4.0",
         curation_reference="https://www.example.com",
         zarr_root_path=zarr_archive,
+        # Competition attributes
+        start_time="2025-01-13T21:59:38Z",
+        end_time="2025-01-20T21:59:38Z",
+        n_classes={"B": 0},
     )
     check_version(competition)
     return competition
